Trying to fix deleting config return to Default

+++ python/lib/ptxprint/gtkview.py
@@ -331,7 +331,7 @@ class GtkViewModel(ViewModel):
         fc = initFontCache()
         lsfonts = self.builder.get_object("ls_font")
 
-        olst = ["b_print", "bx_SavedConfigSettings", "tb_Font", "tb_Layout", "tb_Body", "tb_HeadFoot", "tb_Pictures",
+        olst = ["b_print", "fr_SavedConfigSettings", "tb_Font", "tb_Layout", "tb_Body", "tb_HeadFoot", "tb_Pictures",
                 "tb_Advanced", "tb_Logging", "tb_ViewerEditor", "tb_DiglotBorder"]
         self.initialised = True
         for o in olst:
@@ -405,10 +405,10 @@ class GtkViewModel(ViewModel):
             # for c in ("c_showAdvancedTab", "c_showViewerTab"):
                 # self.builder.get_object(c).set_active(True)
 
-        for c in ("tb_Font", "tb_Advanced", "tb_ViewerEditor", "tb_Tabs", "tb_DiglotBorder", "l_missingPictureString", 
-                  "btn_editPicList", "l_imageTypeOrder", "t_imageTypeOrder", "fr_layoutSpecialBooks", "fr_layoutOther",
+        for c in ("tb_Font", "tb_Advanced", "tb_ViewerEditor", "tb_Tabs", "tb_DiglotBorder", "tb_StyleEdtor", "fr_copyrightLicense",
+                  "l_missingPictureString", "l_imageTypeOrder", "t_imageTypeOrder", "fr_layoutSpecialBooks", "fr_layoutOther",
                   "s_colgutteroffset", "fr_Footer", "bx_TopMarginSettings", "gr_HeaderAdvOptions", "l_colgutteroffset",
-                  "c_usePicList", "c_skipmissingimages", "c_useCustomFolder", "btn_selectFigureFolder", "c_exclusiveFiguresFolder", 
+                  "c_skipmissingimages", "c_useCustomFolder", "btn_selectFigureFolder", "c_exclusiveFiguresFolder", 
                   "c_startOnHalfPage", "c_prettyIntroOutline", "c_marginalverses", "s_columnShift", "c_figplaceholders",
                   "fr_FontConfig", "fr_fallbackFont", "fr_paragraphAdjust", "l_textDirection", "l_colgutteroffset", "fr_hyphenation",
                   "bx_fnCallers", "bx_fnCalleeCaller", "bx_xrCallers", "bx_xrCalleeCaller", "c_fnOverride", "c_xrOverride",
@@ -419,6 +419,7 @@ class GtkViewModel(ViewModel):
                   "c_variableLineSpacing", "c_pagegutter", "s_pagegutter", "fcb_textDirection", "l_digits", "fcb_digits",
                   "t_invisiblePassword", "t_configNotes", "l_notes", "c_elipsizeMissingVerses", "fcb_glossaryMarkupStyle",
                   "gr_fnAdvOptions", "c_figexclwebapp", "bx_horizRule", "l_glossaryMarkupStyle"):
+            # print(c)
             self.builder.get_object(c).set_visible(not self.get("c_hideAdvancedSettings"))
 
         # Resize Main UI Window appropriately
@@ -610,18 +611,25 @@ class GtkViewModel(ViewModel):
         super().writeConfig(cfgname=cfgname)
 
     def onDeleteConfig(self, btn):
-        delCfgPath = self.configPath(cfgname=self.get("t_savedConfig"))
-        if not os.path.exists(os.path.join(delCfgPath, "ptxprint.cfg")):
-            self.doError(_("Internal error occurred, trying to delete a directory tree"), secondary=_("Folder: ")+delCfgPath)
+        cfg = self.get("t_savedConfig")
+        delCfgPath = self.configPath(cfgname=cfg)
+        if cfg == 'Default':
+            self.doError(_("Can't delete 'Default' configuration!"), secondary=_("Folder: ") + delCfgPath)
             return
-        try: # Delete the entire folder
-            rmtree(delCfgPath)
-        except OSError:
-            self.doError(_("Can't delete that configuration from disk"), secondary=_("Folder: ") + delCfgPath)
-        self.updateSavedConfigList()
-        self.set("t_savedConfig", "")
-        self.set("t_configNotes", "")
-        self.updateDialogTitle()
+        else:
+            if not os.path.exists(os.path.join(delCfgPath, "ptxprint.cfg")):
+                self.doError(_("Internal error occurred, trying to delete a directory tree"), secondary=_("Folder: ")+delCfgPath)
+                return
+            try: # Delete the entire folder
+                rmtree(delCfgPath)
+            except OSError:
+                self.doError(_("Can't delete that configuration from disk"), secondary=_("Folder: ") + delCfgPath)
+            self.updateSavedConfigList()
+            self.set("t_savedConfig", "Default")
+            # self.onConfigNameChanged("Default")
+            # self.set("t_configNotes", "")
+            self.loadConfig("Default")
+            self.updateDialogTitle()
 
     def updateBookList(self):
         self.bookNoUpdate = True
@@ -1373,7 +1381,7 @@ class GtkViewModel(ViewModel):
         lockBtn.set_sensitive(False)
         self.updateProjectSettings(None, saveCurrConfig=True, configName="Default")
         self.updateSavedConfigList()
-        for o in ["b_print", "bx_SavedConfigSettings", "tb_Font", "tb_Layout", "tb_Body", "tb_HeadFoot", "tb_Pictures",
+        for o in ["b_print", "fr_SavedConfigSettings", "tb_Font", "tb_Layout", "tb_Body", "tb_HeadFoot", "tb_Pictures",
                   "tb_Advanced", "tb_Logging", "tb_ViewerEditor", "tb_DiglotBorder"]:
             self.builder.get_object(o).set_sensitive(True)
         self.updateFonts()
